feat: update timeline
